factor out more complicate debug helpers into debugext package
