+++ src/x/debug/ext/namespace.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package debug
+package extdebug
 
 import (
 	"bytes"
@@ -31,6 +31,7 @@ import (
 	"github.com/m3db/m3/src/cluster/placementhandler/handleroptions"
 	"github.com/m3db/m3/src/query/api/v1/handler/namespace"
 	"github.com/m3db/m3/src/query/generated/proto/admin"
+	"github.com/m3db/m3/src/x/debug"
 	"github.com/m3db/m3/src/x/instrument"
 	xhttp "github.com/m3db/m3/src/x/net/http"
 
@@ -47,7 +48,7 @@ func NewNamespaceInfoSource(
 	clusterClient clusterclient.Client,
 	allDefaults []handleroptions.ServiceNameAndDefaults,
 	instrumentOpts instrument.Options,
-) (Source, error) {
+) (debug.Source, error) {
 	var (
 		m3dbDefault handleroptions.ServiceNameAndDefaults
 		found       bool
